Code refactoring: split DocumentListEditing integration tests into separate files for better readability.
